Revert "Оголосив змінні в структурі Property, змінив кількість доступних житла на 10 та додав 10 доступних житла до функції allAvailableProperties"
